requirements.txt
Appium-Python-Client==0.46
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
importlib-metadata==0.19
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
packaging==19.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.1
python-dateutil==2.8.0
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.5.2
